requirements.txt
pyqrcode
python-dotenv
pytelegrambotapi
